refactor: split aviation service layer
